@@ -1,5 +1,5 @@
 {
-  "name": "grav-plugin-form",
+  "name": "getgrav/grav-plugin-form",
   "type": "grav-plugin",
   "description": "Grav Plugin Form",
   "keywords": ["form"],
@@ -15,17 +15,26 @@
   ],
   "support": {
     "issues": "https://github.com/getgrav/grav-plugin-form/issues",
-    "irc": "https://gitter.im/getgrav/grav",
-    "forum": "http://getgrav.org/forum",
+    "irc": "https://chat.getgrav.org",
+    "forum": "https://discourse.getgrav.org",
     "docs": "https://github.com/getgrav/grav-plugin-form/blob/master/README.md"
   },
   "require": {
-    "php": ">=5.4.0"
+    "php": ">=7.1.3",
+    "ext-json": "*",
+    "google/recaptcha": "^1.2"
   },
-
   "autoload": {
     "psr-4": {
       "Grav\\Plugin\\Form\\": "classes/"
+    },
+    "classmap": [
+      "form.php"
+    ]
+  },
+  "config": {
+    "platform": {
+      "php": "7.1.3"
     }
   }
 }
