@@ -1,9 +1,11 @@
 <?php
 namespace Grav\Plugin;
 
+use Composer\Autoload\ClassLoader;
 use Grav\Common\Grav;
 use Grav\Common\Page\Page;
 use Grav\Common\Plugin;
+use Grav\Common\Scheduler\Scheduler;
 use Grav\Plugin\TNTSearch\GravTNTSearch;
 use RocketTheme\Toolbox\Event\Event;
 use TeamTNT\TNTSearch\Exceptions\IndexNotFoundException;
@@ -40,10 +42,13 @@ class TNTSearchPlugin extends Plugin
                 ['autoload', 100000],
                 ['onPluginsInitialized', 0]
             ],
-            'onTwigLoader' => ['onTwigLoader', 0],
-            'onTNTSearchReIndex' => ['onTNTSearchReIndex', 0],
-            'onTNTSearchIndex' => ['onTNTSearchIndex', 0],
-            'onTNTSearchQuery' => ['onTNTSearchQuery', 0],
+            'onSchedulerInitialized'    => ['onSchedulerInitialized', 0],
+            'onTwigLoader'              => ['onTwigLoader', 0],
+            'onTNTSearchReIndex'        => ['onTNTSearchReIndex', 0],
+            'onTNTSearchIndex'          => ['onTNTSearchIndex', 0],
+            'onTNTSearchQuery'          => ['onTNTSearchQuery', 0],
+            'onFlexObjectSave'          => ['onObjectSave', 0],
+            'onFlexObjectDelete'        => ['onObjectDelete', 0],
         ];
     }
 
@@ -62,11 +67,9 @@ class TNTSearchPlugin extends Plugin
      */
     public function onPluginsInitialized()
     {
-
-
         if ($this->isAdmin()) {
 
-            $this->grav['tntsearch'] = $this->getSearchObjectType();
+            $this->GravTNTSearch();
             $route = $this->config->get('plugins.admin.route');
             $base = '/' . trim($route, '/');
             $this->admin_route = $this->grav['base_url'] . $base;
@@ -80,8 +83,8 @@ class TNTSearchPlugin extends Plugin
 
             if ($this->config->get('plugins.tntsearch.enable_admin_page_events', true)) {
                 $this->enable([
-                    'onAdminAfterSave' => ['onAdminAfterSave', 0],
-                    'onAdminAfterDelete' => ['onAdminAfterDelete', 0],
+                    'onAdminAfterSave' => ['onObjectSave', 0],
+                    'onAdminAfterDelete' => ['onObjectDelete', 0],
                 ]);
             }
 
@@ -95,11 +98,29 @@ class TNTSearchPlugin extends Plugin
     }
 
     /**
+     * Add index job to Grav Scheduler
+     * Requires Grav 1.6.0 - Scheduler
+     */
+    public function onSchedulerInitialized(Event $e)
+    {
+        if ($this->config->get('plugins.tntsearch.scheduled_index.enabled')) {
+            /** @var Scheduler $scheduler */
+            $scheduler = $e['scheduler'];
+            $at = $this->config->get('plugins.tntsearch.scheduled_index.at');
+            $logs = $this->config->get('plugins.tntsearch.scheduled_index.logs');
+            $job = $scheduler->addFunction('Grav\Plugin\TNTSearchPlugin::indexJob', [], 'tntsearch-index');
+            $job->at($at);
+            $job->output($logs);
+            $job->backlink('/plugins/tntsearch');
+        }
+    }
+
+    /**
      * Function to force a reindex from your own plugins
      */
     public function onTNTSearchReIndex()
     {
-        $this->grav['tntsearch']->createIndex();
+        $this->GravTNTSearch()->createIndex();
     }
 
     /**
@@ -113,7 +134,10 @@ class TNTSearchPlugin extends Plugin
         $fields = $e['fields'];
 
         if ($page && $page instanceof Page && isset($page->header()->author)) {
-            $fields->author = $page->header()->author;
+            $author = $page->header()->author;
+            if (is_string($author)) {
+                $fields->author = $author;
+            }
         }
     }
 
@@ -198,7 +222,7 @@ class TNTSearchPlugin extends Plugin
             }
 
             try {
-                $this->results = $this->grav['tntsearch']->search($this->query);
+                $this->results = $this->GravTNTSearch()->search($this->query);
             } catch (IndexNotFoundException $e) {
                 $this->results = ['number_of_hits' => 0, 'hits' => [], 'execution_time' => 'missing index'];
             }
@@ -267,52 +291,52 @@ class TNTSearchPlugin extends Plugin
 
             // disable warnings
             error_reporting(1);
+            // disable execution time
+            set_time_limit(0);
 
-            // capture content
-            ob_start();
-            $this->grav['tntsearch']->createIndex();
-            ob_get_clean();
-
-            list($status, $msg) = $this->getIndexCount();
+            list($status, $msg, $output) = $this->indexJob();
 
             $json_response = [
                 'status'  => $status ? 'success' : 'error',
-                'message' => '<i class="fa fa-book"></i> ' . $msg
+                'message' => $msg
             ];
+
             echo json_encode($json_response);
             exit;
         }
 
     }
 
-
-
     /**
      * Perform an 'add' or 'update' for index data as needed
      *
      * @param $event
      * @return bool
      */
-    public function onAdminAfterSave($event)
+    public function onObjectSave($event)
     {
-        $obj = $event['object'];
+        $obj = $event['object'] ?: $event['page'];
 
-        $this->grav['tntsearch']->updateIndex($obj);
+        if ($obj) {
+            $this->GravTNTSearch()->updateIndex($obj);
+        }
 
         return true;
     }
 
     /**
-     * Perform an 'add' or 'update' for index data as needed
+     * Perform a 'delete' for index data as needed
      *
      * @param $event
      * @return bool
      */
-    public function onAdminAfterDelete($event)
+    public function onObjectDelete($event)
     {
-        $obj = $event['object'];
+        $obj = $event['object'] ?: $event['page'];
 
-        $this->grav['tntsearch']->deleteIndex($obj);
+        if ($obj) {
+            $this->GravTNTSearch()->deleteIndex($obj);
+        }
 
         return true;
     }
@@ -323,8 +347,9 @@ class TNTSearchPlugin extends Plugin
     public function onTwigAdminVariables()
     {
         $twig = $this->grav['twig'];
+        $gtnt = $this->GravTNTSearch();
 
-        list($status, $msg) = $this->getIndexCount();
+        list($status, $msg) = $this->getIndexCount($gtnt);
 
         if ($status === false) {
             $message = '<i class="fa fa-binoculars"></i> <a href="/'. trim($this->admin_route, '/') . '/plugins/tntsearch">TNTSearch must be indexed before it will function properly.</a>';
@@ -353,14 +378,22 @@ class TNTSearchPlugin extends Plugin
     /**
      * Wrapper to get the number of documents currently indexed
      *
+     * @param $gtnt GravTNTSearch
      * @return array
      */
-    protected function getIndexCount()
+    protected static function getIndexCount($gtnt)
     {
         $status = true;
         try {
-            $this->grav['tntsearch']->selectIndex();
-            $msg = $this->grav['tntsearch']->tnt->totalDocumentsInCollection() . ' documents indexed';
+            $msg = '';
+            $gtnt->selectIndex();
+            $doc_count = $gtnt->tnt->totalDocumentsInCollection();
+
+            $language = Grav::instance()['language'];
+            if ($language->enabled()) {
+                $msg .= 'Processed ' . count($language->getLanguages()) . ' languages, each with ';
+            }
+            $msg .=  $doc_count . ' documents reindexed';
         } catch (IndexNotFoundException $e) {
             $status = false;
             $msg = "Index not created";
@@ -391,5 +424,54 @@ class TNTSearchPlugin extends Plugin
         }
     }
 
+    public static function indexJob()
+    {
+        $grav = Grav::instance();
+
+        $grav['debugger']->enabled(false);
+        $grav['twig']->init();
+
+        $language = $grav['language'];
+
+        ob_start();
+
+        if ($language->enabled()) {
+            foreach ($language->getLanguages() as $lang) {
+                $language->init();
+                $language->setActive($lang);
+
+                echo("\nLanguage: $lang\n");
+                $grav['pages']->init();
+                $gtnt = static::getSearchObjectType();
+                $gtnt->createIndex();
+            }
+        } else {
+            $grav['pages']->init();
+            $gtnt = static::getSearchObjectType();
+            $gtnt->createIndex();
+        }
+        
+        $output = ob_get_clean();
+
+        // Reset and get index count and status
+        $gtnt = static::getSearchObjectType();
+        list($status, $msg) = static::getIndexCount($gtnt);
+
+        return [$status, $msg, $output];
+    }
+
+    /**
+     * Helper to initialize TNTSearch if required
+     *
+     * @return TNTSearch\GravTNTSearch
+     */
+    protected function GravTNTSearch()
+    {
+        if (!isset($this->grav['tntsearch'])) {
+            $this->grav['tntsearch'] = static::getSearchObjectType();
+        }
+
+        return $this->grav['tntsearch'];
+    }
 
 }
